Statement now takes ownership of the query string rather than allowing it to be discarded after parsing. Also refactored some classes (Statement, Portal, BinderSherpa) to have .cpp files to adhere to codebase best practices.
