+++ src/include/network/postgres/portal.h
@@ -21,10 +21,11 @@ namespace terrier::network {
 class Portal {
  public:
   /**
-   * Constructor that doesnt have params or result_formats, i.e. Simple Query protocol
+   * Constructor that doesnt have params or result_formats, i.e. Simple Query protocol. We default the output format to
+   * text for Simple Query protocol per the spec.
    * @param statement statement that this Portal refers to
    */
-  explicit Portal(const common::ManagedPointer<Statement> statement) : Portal(statement, {}, {FieldFormat::text}) {}
+  explicit Portal(common::ManagedPointer<Statement> statement);
 
   /**
    * Constructor that doesnt have params or result_formats, i.e. Extended Query protocol
@@ -32,31 +33,28 @@ class Portal {
    * @param params params for this query
    * @param result_formats output formats for this query
    */
-  Portal(const common::ManagedPointer<Statement> statement, std::vector<type::TransientValue> &&params,
-         std::vector<FieldFormat> &&result_formats)
-      : statement_(statement), params_(std::move(params)), result_formats_(std::move(result_formats)) {}
+  Portal(common::ManagedPointer<Statement> statement, std::vector<type::TransientValue> &&params,
+         std::vector<FieldFormat> &&result_formats);
 
   /**
    * @return Statement that this Portal references
    */
-  common::ManagedPointer<Statement> GetStatement() const { return statement_; }
+  common::ManagedPointer<Statement> GetStatement() const;
 
   /**
    * @return the optimized physical plan for this query
    */
-  common::ManagedPointer<planner::AbstractPlanNode> PhysicalPlan() const { return statement_->PhysicalPlan(); }
+  common::ManagedPointer<planner::AbstractPlanNode> PhysicalPlan() const;
 
   /**
    * @return output formats for this query
    */
-  const std::vector<FieldFormat> &ResultFormats() const { return result_formats_; }
+  const std::vector<FieldFormat> &ResultFormats() const;
 
   /**
    * @return params for this query
    */
-  common::ManagedPointer<const std::vector<type::TransientValue>> Parameters() {
-    return common::ManagedPointer(&params_);
-  }
+  common::ManagedPointer<const std::vector<type::TransientValue>> Parameters();
 
  private:
   const common::ManagedPointer<network::Statement> statement_;
